Refactor queues
Refactor queues
