+++ storage-service/src/app.ts
@@ -1,7 +1,8 @@
 import express from "express";
 import dotenv from "dotenv";
 import imageRoutes from "./routes/imageRoutes";
-import { RabbitMQService } from "./services/rabbitMQService";
+import { UploadQueueService } from "./services/uploadQueueService";
+import ImageProcessingService from "./services/imageProcessingService";
 
 dotenv.config();
 
@@ -15,7 +16,9 @@ app.use("/download", imageRoutes);
 
 app.listen(port, () => {
   console.log(`Server is running on port ${port}`);
-  const rabbitMQService = new RabbitMQService();
+
+  const uploadService = new ImageProcessingService();
+  const rabbitMQService = new UploadQueueService(uploadService);
   rabbitMQService.initialize(); // Initialize RabbitMQ service and listen for messages to upload images
 });
